Issue #424: oc: Remove extract.go
Now that the decompression code moved to the generic extract package,
oc/extract.go is fairly small, we can move it to metadata.go

This is related to https://github.com/code-ready/crc/issues/424
